chore(webflow): restructure typings
