+++ integrations/webflow/definitions/actions/index.ts
@@ -1,0 +1,2 @@
+export * as collections from './collections'
+export * as items from './items'
