--- integrations/webflow/src/collections/collection.ts
@@ -1,6 +1,0 @@
-import { z } from '@botpress/sdk'
-import { CollectionDetailsSchema, CollectionSchema, fieldTypeSchema } from './collectionSchema'
-
-export type fieldType = z.infer<typeof fieldTypeSchema>
-export type CollectionDetails = z.infer<typeof CollectionDetailsSchema>
-export type Collection = z.infer<typeof CollectionSchema>
